moving codgen modules into appropriate locations (#621)
