+++ src/bloqade/emulate/codegen/hamiltonian.py
