+++ src/bloqade/ir/routine/bloqade.py
@@ -25,8 +25,8 @@ class BloqadePythonRoutine(RoutineBase):
     ) -> LocalBatch:
         from bloqade.ir.analysis.assignment_scan import AssignmentScan
         from bloqade.codegen.common.assign_variables import AssignAnalogCircuit
-        from bloqade.emulate.codegen.emulator_ir import EmulatorProgramCodeGen
-        from bloqade.emulate.codegen.rydberg_hamiltonian import CompileCache
+        from bloqade.codegen.emulator_ir import EmulatorProgramCodeGen
+        from bloqade.emulate.codegen.hamiltonian import CompileCache
         from bloqade.task.bloqade import BloqadeTask
 
         circuit, params = self.source.parse_source()
